add countdown timer to newItems
